deploy/requirements.txt
aws-cdk.core==1.47.0
aws_cdk.aws_lambda==1.47.0